feat(script): 支持移除所有gpu分区
1. 可以通过Remove-Adapter -GPUName All移除指定虚拟机中的全部gpu分区

+++ Remove-Adapter.ps1
@@ -4,5 +4,11 @@ param([string]$VMName = $(throw "Parameter missing: -VMName"),
 
 Import-Module $PSScriptRoot/Utility.psm1
 
-$AdapterId = Get-AdapterId -VMName $VMName -GPUName $GPUName
-Remove-VMGpuPartitionAdapter -VMName $VMName -AdapterId $AdapterId
+If ($GPUName -ne "All") {
+    $AdapterId = Get-AdapterId -VMName $VMName -GPUName $GPUName
+    Remove-VMGpuPartitionAdapter -VMName $VMName -AdapterId $AdapterId
+}
+Else {
+    # 移除所有gpu分区
+    Remove-VMGpuPartitionAdapter -VMName $VMName
+}
